% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{format_date}
\alias{format_date}
\title{Format a Date in SAS Date9 format}
\usage{
format_date(date, name)
}
\arguments{
\item{date}{The input date}

\item{name}{the date format. Either "Date9" or "Date7"}
}
\value{
The input date formatted as a string
}
\description{
Format a Date in SAS Date9 format
}
\examples{
dt <- as.Date("1990-01-31")
format_date(dt)
}
